port make custom function easier
