+++ libraries/botbuilder-lg/tests/testData/examples/customFunction2.lg
@@ -1,0 +1,2 @@
+# custom
+- ${contoso.sqrt("16") + contoso.sqrt("4")}
